Move script tests to separate tests.rs module
